follow up pr remarks
