+++ test-utils/src/index.js
@@ -1,2 +1,23 @@
-export * from './act';
-export * from './setupRerender';
+import { Component, options } from 'preact';
+
+/**
+ * Setup a rerender function that will drain the queue of pending renders
+ * @returns {() => void}
+ */
+export function setupRerender() {
+	Component.__test__previousDebounce = options.debounceRendering;
+	options.debounceRendering = cb => Component.__test__drainQueue = cb;
+	return () => Component.__test__drainQueue && Component.__test__drainQueue();
+}
+
+export function act(cb) {
+	const rerender = setupRerender();
+	let flush;
+	options.afterPaint = (fc) => flush = fc;
+	cb();
+	if (flush) {
+		flush();
+	}
+	rerender();
+	options.afterPaint = undefined;
+}
