--- test-utils/src/setupRerender.js
@@ -1,11 +1,0 @@
-import { Component, options } from 'preact';
-
-/**
- * Setup a rerender function that will drain the queue of pending renders
- * @returns {() => void}
- */
-export function setupRerender() {
-	Component.__test__previousDebounce = options.debounceRendering;
-	options.debounceRendering = cb => Component.__test__drainQueue = cb;
-	return () => Component.__test__drainQueue && Component.__test__drainQueue();
-}
